requirements.txt
# POWER-BHOOMI - Karnataka Land Records Search Tool
# ================================================
# Install with: pip install -r requirements.txt

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# HTTP & API
requests>=2.28.0
urllib3>=1.26.0

# Browser Automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# Data Processing
pandas>=2.0.0
beautifulsoup4>=4.12.0

# Enterprise Features
pyyaml>=6.0.0
psutil>=5.9.0

# Jupyter Notebook (optional - for notebook version)
jupyter>=1.0.0
ipykernel>=6.0.0
